+++ src/TypeName.php
@@ -53,21 +53,51 @@ final class TypeName
         $this->simpleName    = $simpleName;
     }
 
-    public function getNamespaceName(): ?string
+    public function namespaceName(): ?string
     {
         return $this->namespaceName;
     }
 
-    public function getSimpleName(): string
+    public function simpleName(): string
     {
         return $this->simpleName;
     }
 
-    public function getQualifiedName(): string
+    public function qualifiedName(): string
     {
         return $this->namespaceName === null
              ? $this->simpleName
              : $this->namespaceName . '\\' . $this->simpleName;
+    }
+
+    /**
+     * @deprecated Use namespaceName() instead
+     *
+     * @codeCoverageIgnore
+     */
+    public function getNamespaceName(): ?string
+    {
+        return $this->namespaceName();
+    }
+
+    /**
+     * @deprecated Use simpleName() instead
+     *
+     * @codeCoverageIgnore
+     */
+    public function getSimpleName(): string
+    {
+        return $this->simpleName();
+    }
+
+    /**
+     * @deprecated Use qualifiedName() instead
+     *
+     * @codeCoverageIgnore
+     */
+    public function getQualifiedName(): string
+    {
+        return $this->qualifiedName();
     }
 
     public function isNamespaced(): bool